Convert to TS, isolate the Swipe.Navigator
Also converts getTabBarHeight() to TAB_BAR_HEIGHT
