+++ src/screens/discover/components/DiscoverSearch.js
@@ -23,7 +23,7 @@ import { ethereumUtils } from '@/utils';
 import { Network } from '@/helpers';
 import { getPoapAndOpenSheetWithQRHash, getPoapAndOpenSheetWithSecretWord } from '@/utils/poaps';
 import { navigateToMintCollection } from '@/resources/reservoir/mints';
-import { getHeaderHeight } from '@/navigation/SwipeNavigator';
+import { TAB_BAR_HEIGHT } from '@/navigation/SwipeNavigator';
 
 export const SearchContainer = styled(Row)({
   height: '100%',
@@ -47,7 +47,7 @@ export default function DiscoverSearch() {
 
   const { colors } = useTheme();
   const profilesEnabled = useExperimentalFlag(PROFILES);
-  const marginBottom = useMemo(() => getHeaderHeight(), []);
+  const marginBottom = TAB_BAR_HEIGHT;
 
   const currencySelectionListRef = useRef();
   const [searchQueryForSearch] = useDebounce(searchQuery, 350);
